+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.37.0
 
-- Emit `hasNextError` when an error occurs when calling `hasNext`. Useful for debugging
+- Emit `cursorError` when an error occurs when calling `hasNext`. Useful for debugging
   syncing issues like a missing oplog entry caused by an inadequate oplog window.
 
 # 0.36.0